requirements.txt
pyppeteer
pillow
numpy
matploylib
opencv-python